refactor(gateway): split session route tests into a sibling file
sessions.rs had grown past the 1000-line PR file-size gate (new file,
1658 lines). Move the test module into sessions_tests.rs, included via
`#[cfg(test)] #[path = "sessions_tests.rs"] mod tests;` — the module
stays a child of `sessions`, so `use super::*` resolves identically and
no test changes. sessions.rs is now 992 lines; tests 666. No logic change.

Refs #972 #973
